普通にdoubleで通るやんけ

+++ POH6/rio/answer.c
@@ -1,20 +1,20 @@
 /*
- * 結果 https://paiza.jp/poh/joshibato/rio/result/338ee274
+ * 結果 https://paiza.jp/poh/joshibato/rio/result/eb379198
  */
 #include <stdio.h>
 #include <math.h>
 
 int main(void) {
     int n, i, t, ans;
-    long double w, c, s, wc;
+    double w, c, s, wc;
     
     scanf("%d", &n);
     
-    w = 0.0L;
-    c = 0.0L;
+    w = 0.0;
+    c = 0.0;
     
     for (i = 0; i < n; i++) {
-        scanf("%d %Lf", &t, &s);
+        scanf("%d %lf", &t, &s);
         switch (t) {
         case 1:
             w += s;
@@ -30,7 +30,7 @@ int main(void) {
         }
     }
     
-    ans = (int)floor((double)(100.0L * c / (w + c)));
+    ans = (int)floor(100.0 * c / (w + c));
     
     printf("%d\n", ans);
